feat: add loading screen to wait access API

+++ public/js/index.js
@@ -7,7 +7,20 @@ $("#signInButton").on("click", function (e) {
     username: email,
     password: password,
   };
-  fetch("http://localhost:5001/api/v1/auth/login", {
+
+  $(this).prop("disabled", true);
+
+  const loadingSwal = Swal.fire({
+    title: "Please wait...",
+    allowOutsideClick: false,
+    allowEscapeKey: false,
+    showConfirmButton: false,
+    onBeforeOpen: () => {
+      Swal.showLoading();
+    },
+  });
+
+  fetch("/api/v1/auth/login", {
     method: "POST",
     headers: {
       "Content-Type": "application/json",
@@ -20,17 +33,24 @@ $("#signInButton").on("click", function (e) {
     .then((data) => {
       console.log("isi data", data);
       if (data.success === true) {
+        loadingSwal.close();
+
         Swal.fire({
           title: "Success Login",
-          text: "Successfully logged in, you will be redirected to dashboard in 3 seconds..",
+          text: "Successfully logged in, you will be redirected to the dashboard in 3 seconds..",
           icon: "success",
           confirmButtonText: "OK",
+        }).then((result) => {
+          if (result.isConfirmed || result.isDismissed) {
+            localStorage.setItem("userToken", JSON.stringify(data.data));
+            setTimeout(() => {
+              window.location.href = "dashboard.html";
+            }, 3000);
+          }
         });
-        localStorage.setItem("userToken", JSON.stringify(data.data));
-        setTimeout(() => {
-          window.location.href = "dashboard.html";
-        }, 3000);
       } else {
+        loadingSwal.close();
+
         Swal.fire({
           title: "Failed Login",
           text: data.message,
@@ -41,11 +61,92 @@ $("#signInButton").on("click", function (e) {
     })
     .catch((error) => {
       console.error("Error:", error.message);
+      loadingSwal.close();
+
       Swal.fire({
         title: "Internal Server Error",
         text: `${error.message}. Please contact the admin.`,
         icon: "error",
         confirmButtonText: "OK",
       });
+    })
+    .finally(() => {
+      $(this).prop("disabled", false);
+    });
+});
+
+$("#signInButton").on("click", function (e) {
+  e.preventDefault();
+  $(this).prop("disabled", true);
+
+  const loadingSwal = Swal.fire({
+    title: "Please wait...",
+    allowOutsideClick: false,
+    allowEscapeKey: false,
+    showConfirmButton: false,
+    onBeforeOpen: () => {
+      Swal.showLoading();
+    },
+  });
+
+  const username = document.getElementById("username").value;
+  const password = document.getElementById("password").value;
+  const gender = document.getElementById("gender").value;
+
+  const loginData = {
+    username: username,
+    password: password,
+    gender: gender,
+  };
+
+  fetch("/api/v1/auth/register", {
+    method: "POST",
+    headers: {
+      "Content-Type": "application/json",
+    },
+    body: JSON.stringify(loginData),
+  })
+    .then((response) => {
+      return response.json();
+    })
+    .then((data) => {
+      console.log("isi data", data);
+      if (data.success === true) {
+        loadingSwal.close();
+
+        Swal.fire({
+          title: "Success Register",
+          text: "Successfully created an account. You may now log in to the application.",
+          icon: "success",
+          confirmButtonText: "OK",
+        }).then((result) => {
+          if (result.isConfirmed || result.isDismissed) {
+            window.location.href = "index.html";
+          }
+        });
+      } else {
+        loadingSwal.close();
+
+        Swal.fire({
+          title: "Failed Register",
+          text: data.message,
+          icon: "error",
+          confirmButtonText: "OK",
+        });
+      }
+    })
+    .catch((error) => {
+      console.error("Error:", error);
+      loadingSwal.close();
+
+      Swal.fire({
+        title: "Internal Server Error",
+        text: `${error.message}. Please contact the admin.`,
+        icon: "error",
+        confirmButtonText: "OK",
+      });
+    })
+    .finally(() => {
+      $(this).prop("disabled", false);
     });
 });
